+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,52 +1,5 @@
+import { authOptions } from "@/utils/auth";
 import NextAuth from "next-auth/next";
-import GoogleProvider from "next-auth/providers/google";
-import GithubProvider from "next-auth/providers/github";
-import NaverProvider from "next-auth/providers/naver";
-import KakaoProvider from "next-auth/providers/kakao";
-import { NextAuthOptions } from "next-auth";
-
-export const authOptions: NextAuthOptions = {
-  // Configure one or more authentication providers
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID || "",
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
-    }),
-    GithubProvider({
-      clientId: process.env.GITHUB_ID || "",
-      clientSecret: process.env.GITHUB_SECRET || "",
-    }),
-    NaverProvider({
-      clientId: process.env.NAVER_CLIENT_ID || "",
-      clientSecret: process.env.NAVER_CLIENT_SECRET || "",
-    }),
-    KakaoProvider({
-      clientId: process.env.KAKAO_CLIENT_ID || "",
-      clientSecret: process.env.KAKAO_CLIENT_SECRET || "",
-    }),
-  ],
-//   debug: true,
-  callbacks: {
-    async session({ session, token }) {
-      // Send properties to the client, like an access_token and user id from a provider.
-       const user = session?.user;
-      if(user) {
-        session.user = {
-          ... user,
-          name: user.name ? user.name : user?.email?.split('@')[0],
-        }
-      }
-      console.log('session', session.user);
-      return session;
-    },
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-      }
-      return token;
-    },
-  },
-};
 
 const handler = NextAuth(authOptions);
