+++ src/components/comments/Comments.tsx
@@ -9,6 +9,7 @@ type Prop = {
 export default function Comments({ slug }: Prop) {
   return (
     <div>
+      {/* @ts-expect-error Async Server Component */}
       <CommentInput slug={slug} />
       <Suspense fallback={<div>댓글을 불러오는 중입니다.</div>}>
         {/* @ts-expect-error Async Server Component */}
